add subcmd: checksum for apis
